Parser Rewrite (MathML)
