+++ src/parserold.js
@@ -1,0 +1,245 @@
+'use strict';
+
+import { greek } from './calculator.js'
+
+// func[default]{argc}: [$0]{$1}{$2}...{$argc}
+const command_subs = {
+	...Object.fromEntries(greek.map(x => [x, x])),
+	'sqrt[2]{1}': 'root($1, $0)',
+	'frac{2}': '(($1)/($2))',
+	'operatorname{1}': ' $1 ',
+	'left': '',
+	'right': '',
+	'cdot': '*',
+	'text': '',
+	'infty': 'infinity',
+	'%': '%'
+}
+
+const TOK_EXPRESSION = 'tok-expression';
+const TOK_COMMAND = 'tok-command';
+const TOK_UNDERSCORE = 'tok-underscore';
+const TOK_PUCNT = 'tok-punct';
+/**
+ * 
+ * @param {String} str
+ */
+const tokenize_latex = str => {
+	const tokens = [];
+	while (str.length) {
+		let c = str[0];
+		let body;
+		let type;
+
+		do {
+			if ('[]{}()'.includes(c)) {
+				type = TOK_PUCNT;
+				body = c;
+				break;
+			}
+
+			if (c == '\\') {
+				type = TOK_COMMAND;
+				body = str.match(/^\\(?:[a-zA-Z&*]*[a-zA-Z&*]|[ %])/);
+				if (body == null) {
+					console.error('Unable to parse expression:', str);
+				}
+				body = body[0];
+				break;
+			}
+
+			if (c == '_') {
+				type = TOK_UNDERSCORE;
+				if (str[1] == '{') {
+					body = '_';
+					break;
+				}
+				body = '_' + str[1];
+				break;
+			}
+
+			type = TOK_EXPRESSION;
+			body = str.match(/^[^\\()[\]{}_]+/);
+			if (body == null) {
+				console.error('Unable to parse expression:', str);
+			}
+			body = body[0];
+			break;
+		} while (0);
+
+		str = str.slice(body.length);
+		tokens.push({type, body});
+	}
+
+	return tokens;
+}
+
+const flatten = arr => {
+	if (!(arr instanceof Array)) { return arr; }
+	return arr.map(x => flatten(x)).join('');
+}
+
+
+// Key assumption:
+// Assume that latex is always well formed (because it is generated by MQ)
+// Though the formulas themselves may not be correct
+export const parse_latex = latex => {
+	const tokens = tokenize_latex(latex);
+	let i = 0;
+	
+	const command_map = {};
+	Object.keys(command_subs).forEach(proto => {
+		const cmd = '\\' + proto.match(/(?:[a-zA-Z0-9-&*]*[a-zA-Z-&*]|[ %])/)[0];
+		const sub = command_subs[proto];
+
+		const arg_proto = proto.slice(cmd.length-1);
+		if (arg_proto.length == 0) {
+			// No arguments
+			command_map[cmd] = () => sub;
+			return;
+		}
+
+		let match = arg_proto.match(/^(?:\[(.+?)\])?{(.+?)}$/);
+		const default_arg = match[1];
+		const argc = match[2];
+
+		command_map[cmd] = () => {
+			const parsed = parse_function(argc)
+			const opt = parsed.opt ?? default_arg;
+			let subbed = sub;
+			subbed = subbed.replaceAll('$0', flatten(opt));
+			for (let j = 1; j <= argc; j++) {
+				subbed = subbed.replaceAll(`$${j}`, flatten(parsed.args[j-1]));
+			}
+			return subbed;
+		}
+	})
+
+	// \func[opt]{1}{2}{3}... (Assume max 1 optional argument)
+	const parse_function = (argc) => {
+		i++;
+		const args = [];
+		let opt = null;
+		if (tokens[i].body == '[') {
+			i++;
+			opt = parse_expression();
+			i += 2;
+		} else {
+			i++;
+		}
+
+		for (let j = 0; j < argc; j++) {
+			args[j] = parse_expression();
+			i += 2;
+		}
+		i -= 2;
+
+		return {args, opt};
+	}
+
+	// \func_a^bf(x)
+	// \func_a^{bb}f(x)
+	// \func_{aa}^bf(x)
+	// \func_{aa}^b(f(x))
+	// \func_{aa}^{bb}f(x)
+	const parse_bounded_function = () => {
+		let subscript, superscript, body;
+		i++;
+		if (tokens[i].body != '_') {
+			const tok = tokens[i].body;
+			subscript = tok[1];
+			if (tok.length == 3) {
+				// Case 2
+				i += 2;
+				superscript = parse_expression(true);
+				i++;
+				body = parse_expression(true);
+			} else {
+				// Case 1
+				superscript = tok[3];
+				body = tok.slice(4);
+			}
+		} else {
+			i += 2;
+			subscript = parse_expression(true);
+			i++;
+			if (tokens[i].body == '^') {
+				// Case 4
+				i += 2;
+				superscript = parse_expression(true);
+				i++;
+				body = parse_expression(true);
+			} else {
+				// Case 3
+				const tok = tokens[i].body;
+				superscript = tok[1];
+				body = tok.slice(2);
+			}
+		}
+
+		return {subscript, superscript, body};
+	}
+
+	const parse_expression = lazy => {
+		lazy ??= false;
+		let stack = 0;
+		let terms = [];
+		for (; i < tokens.length; i++) {
+			const tok = tokens[i];
+			if (tok.type == TOK_PUCNT) {
+				if ('[{('.includes(tok.body)) {
+					stack++;
+					terms.push('(');
+				}
+				else {
+					stack--;
+					if (stack >= 0) { terms.push(')'); }
+				}
+				if (stack < 0) { break; }
+				if (lazy && stack == 0) { break; }
+				continue;
+			}
+			if (tok.type == TOK_EXPRESSION) {
+				terms.push(tok.body);
+				continue;
+			}
+
+			if (tok.type == TOK_UNDERSCORE) {
+				let body = tok.body;
+				if (body == '_') {
+					i++;
+					body += parse_expression(true);
+					body = body.replaceAll(' ', '').replaceAll('(', '').replaceAll(')', '');
+				}
+				terms.push(body)
+				continue;
+			}
+				
+			// tok.type == TOK_COMMAND
+			const replacement = command_map[tok.body];
+			if (replacement != undefined) {
+				terms.push(replacement());
+				continue;
+			}
+			
+			switch (tok.body) {
+				case '\\sum': {
+					const {subscript, superscript, body} = parse_bounded_function();
+					const iter_var = subscript.split('=', 1)[0]
+					const lower = subscript.slice(iter_var.length + 1);
+					const upper = superscript;
+					terms.push(`sum(${body}, ${lower}, ${upper}, ${iter_var})`);
+				} break;
+				default: {
+					terms.push(tok.body.slice(1));
+				} break;
+			}
+			if (lazy && stack == 0) { break; }
+		}
+
+		return flatten(terms);
+	}
+	
+	const x = parse_expression();
+	return x;
+}
